Refactor FlutterWhisperKit structure and enhance exports
- Moved the main implementation of the FlutterWhisperKit class to a new file, `flutter_whisper_kit.dart`, for better organization.
- Updated exports in `flutter_whisper_kit.dart` to include necessary model and error handling files.
- Cleaned up the previous `flutter_whisper_kit.dart` file by removing redundant code and ensuring a streamlined structure.
- Adjusted imports in `model_loader.dart` to reflect the new file organization.
