+++ packages/flutter_whisperkit/lib/src/models.dart
@@ -1,10 +1,5 @@
-library flutter_whisperkit.models;
-
 // Export the model classes
 export 'models/decoding_options.dart';
 export 'models/transcription_result.dart';
 export 'models/word_timing.dart';
 export 'models/progress.dart';
-
-// Export error handling
-export 'whisper_kit_error.dart';
